requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
distro==1.9.0
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.16.0
frozenlist==1.7.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
json_repair==0.46.2
kombu==5.5.4
multidict==6.4.4
numpy==2.3.0
openai==1.88.0
packaging==25.0
pgvector==0.4.1
pluggy==1.6.0
postgrest==1.0.2
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.3
redis==6.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
yarl==1.20.1
yt-dlp==2025.6.9
